preview state should be controlled by tableMetadata reducer
