+++ amundsen_application/static/js/containers/TableDetail/DataPreviewButton/index.tsx
@@ -2,14 +2,14 @@ import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 
 import { GlobalState } from "../../../ducks/rootReducer";
-import { getPreviewData } from '../../../ducks/preview/reducer';
+import { getPreviewData } from '../../../ducks/tableMetadata/reducer';
 
 import DataPreviewButton, { ComponentProps, DispatchFromProps, StateFromProps, getStatusFromCode } from '../../../components/TableDetail/DataPreviewButton';
 
 export const mapStateToProps = (state: GlobalState) => {
   return {
-    previewData: state.preview.previewData,
-    status: getStatusFromCode(state.preview.status),
+    previewData: state.tableMetadata.preview.data,
+    status: getStatusFromCode(state.tableMetadata.preview.status),
   };
 };
