--- amundsen_application/static/js/ducks/api/preview/v0.ts
@@ -1,17 +1,0 @@
-import axios from 'axios';
-
-import { GetPreviewDataRequest } from '../../preview/reducer';
-
-export function getPreviewData(action: GetPreviewDataRequest) {
-  return axios({
-    url: '/api/preview/v0/',
-    method: 'POST',
-    data: action.queryParams,
-  })
-  .then((response) => {
-    return { previewData: response.data.previewData, status: response.status };
-  })
-  .catch((error) => {
-    return { previewData: error.response.data.previewData, status: error.response.status };
-  });
-}
