+++ amundsen_application/static/js/ducks/rootReducer.ts
@@ -3,7 +3,6 @@ import { combineReducers } from 'redux';
 import announcements, { AnnouncementsReducerState } from './announcements/reducer';
 import feedback, { FeedbackReducerState } from './feedback/reducer';
 import popularTables, { PopularTablesReducerState } from './popularTables/reducer';
-import preview, { PreviewDataReducerState } from './preview/reducer';
 import search, { SearchReducerState } from './search/reducer';
 import tableMetadata, { TableMetadataReducerState } from './tableMetadata/reducer';
 import tags, { TagReducerState } from './tags/reducer';
@@ -13,7 +12,6 @@ export interface GlobalState {
   announcements: AnnouncementsReducerState;
   feedback: FeedbackReducerState;
   popularTables: PopularTablesReducerState;
-  preview: PreviewDataReducerState;
   search: SearchReducerState;
   tableMetadata: TableMetadataReducerState;
   tags: TagReducerState;
@@ -24,7 +22,6 @@ export default combineReducers<GlobalState>({
   announcements,
   feedback,
   popularTables,
-  preview,
   search,
   tableMetadata,
   tags,
